+++ inject-thrift-client-http-mapper/src/test/scala/com/twitter/finatra/multiserver/Add1HttpServer/Add1Controller.scala
@@ -3,10 +3,9 @@ package com.twitter.finatra.multiserver.Add1HttpServer
 import com.twitter.adder.thriftscala.Adder
 import com.twitter.finagle.http.Request
 import com.twitter.finatra.http.Controller
-import com.twitter.util.Future
 import javax.inject.Inject
 
-class Add1Controller @Inject() (adder: Adder[Future]) extends Controller {
+class Add1Controller @Inject() (adder: Adder.MethodPerEndpoint) extends Controller {
 
   get("/add1") { request: Request =>
     val num = request.getIntParam("num")