+++ inject/inject-thrift-client/src/test/scala/com/twitter/inject/thrift/integration/filtered/ThriftClientFilter.scala
@@ -1,6 +1,0 @@
-package com.twitter.inject.thrift.integration.filtered
-
-import com.twitter.finagle.SimpleFilter
-import com.twitter.scrooge.ThriftStruct
-
-trait ThriftClientFilter[Req <: ThriftStruct, Rep] extends SimpleFilter[Req, Rep]